RIRs looking good!

+++ RoomAcoustiCpp/source/Spatialiser/Main.cpp
@@ -225,7 +225,7 @@ extern "C"
 	*/
 	EXPORT void API RACUpdateReverbTime(const float* T60Data)
 	{
-		Coefficients T60 = Coefficients<>(static_cast<size_t>(NUM_FREQUENCY_BANDS));
+		Coefficients T60 = Coefficients<>(NUM_FREQUENCY_BANDS);
 		for (int i = 0; i < NUM_FREQUENCY_BANDS; i++)
 			T60[i] = static_cast<Real>(T60Data[i]);
